pyproject.toml
[tool.poetry]
name = "python-fastapi-app"
version = "0.1.0"
description = ""
authors = ["Mr. B <bijjubhattaram@gmail.com>"]
readme = "README.md"
packages = [{include = "python_fastapi_app"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "0.95.1"
hypercorn = "0.14.3"
trio = "0.22.0"
toml = "0.10.2"
PyYAML = "6.0"
uvicorn = "0.20.0"
pyodbc = "4.0.35"
starlette = "0.26.1"
requests = "2.28.2"
pydantic = "1.10.4"
http-exceptions = "0.2.1"
PyJWT = "2.6.0"

[tool.poetry.dev-dependencies]
pytest =  "7.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
